+++ utils/run-lifecycle/run-lifecycle.js
@@ -2,23 +2,59 @@
 
 const log = require("libnpm/log");
 const runScript = require("libnpm/run-script");
+const figgyPudding = require("figgy-pudding");
 const npmConf = require("@lerna/npm-conf");
 
 module.exports = runLifecycle;
 module.exports.createRunner = createRunner;
 
-function runLifecycle(pkg, stage, opts) {
-  log.silly("run-lifecycle", stage, pkg.name);
+const LifecycleConfig = figgyPudding(
+  {
+    log: { default: log },
+    // provide aliases for some dash-cased props
+    "ignore-prepublish": {},
+    ignorePrepublish: "ignore-prepublish",
+    "ignore-scripts": {},
+    ignoreScripts: "ignore-scripts",
+    "node-options": {},
+    nodeOptions: "node-options",
+    "script-shell": {},
+    scriptShell: "script-shell",
+    "scripts-prepend-node-path": {},
+    scriptsPrependNodePath: "scripts-prepend-node-path",
+    "unsafe-perm": {},
+    unsafePerm: "unsafe-perm",
+  },
+  {
+    other() {
+      // open up the pudding
+      return true;
+    },
+  }
+);
 
-  const config = {};
+function runLifecycle(pkg, stage, _opts) {
+  // back-compat for @lerna/npm-conf instances
+  // https://github.com/isaacs/proto-list/blob/27764cd/proto-list.js#L14
+  if ("root" in _opts) {
+    // eslint-disable-next-line no-param-reassign
+    _opts = _opts.snapshot;
+  }
+
+  const opts = LifecycleConfig(_opts);
   const dir = pkg.location;
+  const config = {};
 
-  // https://github.com/isaacs/proto-list/blob/27764cd/proto-list.js#L29
-  for (const key of opts.keys) {
-    const val = opts.get(key);
+  if (!pkg.scripts || !pkg.scripts[stage]) {
+    opts.log.silly("run-lifecycle", "No script for %j in %j, continuing", stage, pkg.name);
 
-    // omit falsy and Stream-based (circular) values
-    if (val != null && key !== "logstream" && key !== "log") {
+    return Promise.resolve(pkg);
+  }
+
+  // https://github.com/zkat/figgy-pudding/blob/7d68bd3/index.js#L42-L64
+  for (const [key, val] of opts) {
+    // omit falsy values and circular objects
+    if (val != null && key !== "log" && key !== "logstream") {
       config[key] = val;
     }
   }
@@ -29,12 +65,29 @@ function runLifecycle(pkg, stage, opts) {
   // TODO: remove pkg._id when npm-lifecycle no longer relies on it
   pkg._id = `${pkg.name}@${pkg.version}`; // eslint-disable-line
 
+  // bring along camelCased aliases
+  const {
+    ignorePrepublish,
+    ignoreScripts,
+    nodeOptions,
+    scriptShell,
+    scriptsPrependNodePath,
+    unsafePerm,
+  } = opts;
+
+  opts.log.silly("run-lifecycle", stage, pkg.name);
+
   return runScript(pkg, stage, dir, {
     config,
     dir,
     failOk: false,
-    log,
-    unsafePerm: true,
+    log: opts.log,
+    ignorePrepublish,
+    ignoreScripts,
+    nodeOptions,
+    scriptShell,
+    scriptsPrependNodePath,
+    unsafePerm,
   }).then(
     () => pkg,
     err => {
@@ -58,13 +111,7 @@ function runLifecycle(pkg, stage, opts) {
 }
 
 function createRunner(commandOptions) {
-  const cfg = npmConf(commandOptions);
+  const cfg = npmConf(commandOptions).snapshot;
 
-  return (pkg, stage) => {
-    if (pkg.scripts && pkg.scripts[stage]) {
-      return runLifecycle(pkg, stage, cfg);
-    }
-
-    return Promise.resolve(pkg);
-  };
+  return (pkg, stage) => runLifecycle(pkg, stage, cfg);
 }